+++ app/Models/Image.php
@@ -34,6 +34,19 @@ use Spatie\Activitylog\Models\Activity;
 use Spatie\Activitylog\Traits\LogsActivity;
 use Spatie\Translatable\HasTranslations;
 
+/**
+ * @property int $id
+ * @property int|null $gallery_id
+ * @property string $image
+ * @property int $origin
+ * @property int $copyright
+ * @property string|null $url
+ * @property string|null $alt
+ * @property int $position
+ * @property bool $is_translated
+ * @property Source|null $originImage
+ * @property Source|null $copyrightImage
+ */
 class Image extends Model implements HasComments
 {
     use HasFactory, HasTranslations, LogsActivity, SoftDeletes;
@@ -49,20 +62,16 @@ class Image extends Model implements HasComments
 
     /**
      * Get image origin
-     *
-     * @return BelongsTo
      */
-    public function originImage()
+    public function originImage(): BelongsTo
     {
         return $this->belongsTo(Source::class, 'origin');
     }
 
     /**
      * Get image copyright
-     *
-     * @return BelongsTo
      */
-    public function copyrightImage()
+    public function copyrightImage(): BelongsTo
     {
         return $this->belongsTo(Source::class, 'copyright');
     }