+++ database/factories/GalleryFactory.php
@@ -1,0 +1,43 @@
+<?php
+
+/**
+crowdCuratio - Curating together virtually
+Copyright (C)2026 - berlinHistory e.V.
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program in the file LICENSE.
+
+If not, see <https://www.gnu.org/licenses/>.
+ */
+
+namespace Database\Factories;
+
+use App\Models\Gallery;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+/**
+ * @extends Factory<Gallery>
+ */
+class GalleryFactory extends Factory
+{
+    protected $model = Gallery::class;
+
+    public function definition(): array
+    {
+        return [
+            'title' => $this->faker->sentence(2),
+            'subtitle' => $this->faker->sentence(3),
+            'description' => $this->faker->paragraph(),
+        ];
+    }
+}