+++ database/factories/SourceFactory.php
@@ -1,0 +1,52 @@
+<?php
+
+/**
+crowdCuratio - Curating together virtually
+Copyright (C)2026 - berlinHistory e.V.
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program in the file LICENSE.
+
+If not, see <https://www.gnu.org/licenses/>.
+ */
+
+namespace Database\Factories;
+
+use App\Models\Source;
+use Illuminate\Database\Eloquent\Factories\Factory;
+
+/**
+ * @extends Factory<Source>
+ */
+class SourceFactory extends Factory
+{
+    protected $model = Source::class;
+
+    public function definition(): array
+    {
+        return [
+            'name' => $this->faker->sentence(3),
+            'type' => $this->faker->randomElement(['Origin', 'Copyright']),
+        ];
+    }
+
+    public function origin(): self
+    {
+        return $this->state(['type' => 'Origin']);
+    }
+
+    public function copyright(): self
+    {
+        return $this->state(['type' => 'Copyright']);
+    }
+}